Update exit_prob for the last node in a path
The path-based Infomap function fails if there are any nodes that only appear at the ends of paths (sinks). These nodes received neither any exit_prob nor any within_prob, leading to log-of-zero and divide-by-zero errors later in the function. The added line of code assumes that walkers at the end of a path will jump with 100% probability to a different module -- this isn't quite true, I suppose, but the rest of this function is deterministic as well. Adding probabilistic jumps would require keeping track of all module in-strengths and adjusting all the probabilities accordingly. If you'd like I could implement this, adding an "alpha" parameter link as in PageRank. Setting the default to 1 would preserve the default behavior of the function as it is now. Let me know :)
Thank you to Roshni Padhi for noticing the error.